Fix variables to use ansible syntax
A number of variables in the run and fetch playbook were defined using
shell $varname syntax. This was causing problems with the playbook since
those strings were treated a literal values on the target causing the
playbooks to fail.

Also aligned the fetch of the md5 and tarball targets to reference the
same local directory variable.